refactor(cli): extract compare.py export helpers to _compare_export
Moves the three export functions (_export_csv 33 LOC, _export_json
41 LOC, _export_markdown 111 LOC) out of compare.py into
amx/cli_support/_compare_export.py.

Each exporter pulls the compare-internal collectors and constants
(_collect_run_summary_rows, _aggregate_for_run, _AGGREGATE_COLUMNS,
etc.) via lazy imports tailored to what it actually uses — that
avoids the obvious circular import while keeping each function's
dependency footprint visible.

compare.py re-exports the three names so existing test imports
(from amx.cli_support.commands.compare import _export_csv, ...) keep
working unchanged.

compare.py: 2446 -> 2255 LOC (-191, -8%).
